backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pymongo>=4.6.0,<4.12
motor>=3.3.0,<3.7
beanie>=1.23.0,<2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
Pillow>=10.1.0
aiofiles>=23.2.1
email-validator>=2.1.0
httpx>=0.25.0
